refactor: move todo component out from App.tsx
